fix option box bugs

+++ app.py
@@ -27,7 +27,7 @@ def runpy(f,*options):
     #         return subprocess.check_output([relPath(tempf),*options]).decode('utf-8')
     if getattr(sys, 'frozen', False):
         # The application is frozen
-        f = re.sub('.py','.exe')
+        f = re.sub('.py','.exe',f)
     return subprocess.check_output([sys.executable, relPath(f),*options]).decode('utf-8')
 def profiles():
     return os.listdir(relPath('profiles'))
@@ -85,6 +85,9 @@ def label(profile,image_path,shouldPrint=False,shouldParse=True):
 # =============================== app-specific stuff ===========================
 
 def createProfile(name):
+    if name == '':
+        print('warning: tried to create profile with empty string name')
+        return None
     name = re.sub(' ','_',name)
     match = re.sub(r'\w','',name)
     if match != '':
@@ -103,13 +106,16 @@ def createProfile(name):
     except FileExistsError:
         app.errorBox('profile already exists error','It appears a profile with that name already exists',parent=None)
 def removeProfile(profile):
+    if profile == '':
+        print('warning: tried to remove profile with name empty string.')
+        return None
     shutil.rmtree(relPath('profiles/'+profile))
 def updateOptionBoxes():
     '''updates all option boxes and list boxes after the profiles have been edited'''
+    updateViewListBox()
     updateUseOptionBox()
     updateTrainOptionBox()
     updateRemoveOptionBox()
-    updateViewListBox()
 
 #=================================== press =====================================
 
@@ -121,9 +127,21 @@ def press(button):
         if profile:
             createProfile(profile)
     elif button == 'view profiles':
-        app.showSubWindow('view profiles window')
+        if len(profiles()) > 0:
+            app.showSubWindow('view profiles window')
+        else:
+            try:
+                app.errorBox('no profiles to view error',"To view your profiles, you need to have profiles. Try adding one")
+            except AttributeError:
+                pass
     elif button == 'remove a profile':
-        app.showSubWindow('remove profiles window')
+        if len(profiles()) > 0:
+            app.showSubWindow('remove profiles window')
+        else:
+            try:
+                app.errorBox('no profiles to remove error','There are no profiles to remove. Try adding one')
+            except AttributeError:
+                pass
     elif button == 'train a profile':
         if len(profiles()) > 0:
             app.showSubWindow('train profile window')
@@ -137,7 +155,7 @@ def press(button):
             app.showSubWindow('use profile window')
         else:
             try:
-                app.errorBox('no profiles to use error','To use a profile, you need a profile. Try adding one')
+                app.errorBox('no profiles to use error',"To use a profile, you need a trained profile. Try training one. If you don't have any, try adding one")
             except AttributeError:
                 pass
     elif button == 'choose image directory':
@@ -165,6 +183,8 @@ def press(button):
             app.threadCallback(label,whenDone,profile,image_path,shouldPrint=True)#TODO change to false when done
     elif button == 'train':
         profile = app.getOptionBox('train profiles option box')
+        if profile == '':
+            return None
         profile = re.sub(' (.*)','',profile)
         image_dir = app.getLabel('image_dir')
 
@@ -184,8 +204,11 @@ def press(button):
             app.threadCallback(train,whenDone,profile,image_dir,shouldPrint=True)
     elif button == 'remove':
         profile = app.getOptionBox('remove profiles option box')
-        profile = re.sub(' (.*)','',profile)
-        removeProfile(profile)
+        if profile == '':
+            return None
+        if profile:
+            profile = re.sub(' (.*)','',profile)
+            removeProfile(profile)
         updateOptionBoxes()#TODO thread callback for deleting trained profiles
 
 #============================== main window ====================================
@@ -203,9 +226,12 @@ app.addButton('remove a profile',press)
 app.startSubWindow('train profile window',title="train",modal=True)
 app.startLabelFrame('select profile to train')
 def updateTrainOptionBox():
-    app.changeOptionBox('train profiles option box',profiles())
+    arr = profiles()
+    print(arr)
+    if not arr:
+        arr = ['']
+    app.changeOptionBox('train profiles option box',arr)
 app.addOptionBox('train profiles option box',profiles())
-updateTrainOptionBox()
 app.stopLabelFrame()
 app.setSize(defaultSize)
 app.addButton('choose image directory',press)
@@ -217,7 +243,10 @@ app.stopSubWindow()
 app.startSubWindow('use profile window',title='use',modal=True)
 app.startLabelFrame('select profile to use')
 def updateUseOptionBox():
-    app.changeOptionBox('use profiles option box',trainedProfiles())
+    arr = trainedProfiles()
+    if not arr:
+        arr = ['']
+    app.changeOptionBox('use profiles option box',arr)
 app.addOptionBox('use profiles option box',trainedProfiles())
 app.stopLabelFrame()
 app.addButton('select an image to be labeled',press)
@@ -232,7 +261,10 @@ app.setSize(defaultSize)
 app.setPadding([20,20])
 app.startScrollPane('view profiles scroll pane')
 def updateViewListBox():
-    app.updateListBox('view profiles list box',labeledProfiles())
+    arr = labeledProfiles()
+    if not arr:
+        arr = ['']
+    app.updateListBox('view profiles list box',arr)
 app.addListBox('view profiles list box',labeledProfiles())
 app.stopScrollPane()
 app.stopSubWindow()
@@ -243,8 +275,16 @@ app.startSubWindow('remove profiles window',title='remove',modal=True)
 app.setSize(defaultSize)
 app.startLabelFrame('select profile to remove')
 def updateRemoveOptionBox():
-    app.changeOptionBox('remove profiles option box',profiles())
-app.addOptionBox('remove profiles option box',profiles())
+    arr = profiles()
+    print(arr)
+    if not arr:
+        arr = ['']
+    print(arr)
+    app.changeOptionBox('remove profiles option box',arr)
+arr = profiles()
+if not arr:
+    arr = ['']
+app.addOptionBox('remove profiles option box',arr)
 app.stopLabelFrame()
 app.addButton('remove',press)
 app.stopSubWindow()
